+++ CHANGES.rst
@@ -4,6 +4,7 @@
   number since this is a non-breaking incompatible change.
 * JsonDict now uses JSONB on PostgreSQL 9.4.
 * New CoordinatesMixin adds latitude and longitude columns.
+* Rudimentary NLP methods
 
 0.4.3
 -----